+++ bottery/conf/__init__.py
@@ -4,7 +4,6 @@ from copy import deepcopy
 from importlib import import_module
 
 from bottery.conf import global_settings
-from bottery.exceptions import ImproperlyConfigured
 
 
 class Settings:
@@ -20,10 +19,6 @@ class Settings:
 
     def _configure_settings_path(self):
         base = os.getcwd()
-        settings_path = os.path.join(base, 'settings.py')
-        if not os.path.isfile(settings_path):
-            raise ImproperlyConfigured('Could not find settings module')
-
         sys.path.insert(0, base)
 
     def _import_settings(self):